Add Giphy command
Connects Jeeves to https://github.com/Giphy/GiphyAPI (/random enpoint)

Example call

    !!giphy Chuck Norris

should post the image url of a random gif of Chuck Norris to the chat.

By default images are limited to PG-13 as per the ToS of Stack Exchange.
The plugin uses the default Public Beta key of Giphy, which is rate
limited. A dedicated API key (and rating) can be ctor injected.